FIX: Consul agents have conflicting IDs
When we create an AMI with packer, the consul agent is installed on
 it and creates a unique ID during start.
This is a problem, as we later launch an AMI of a type (apache2 or mysql) many times.
IDs are then the same on different nodes and would conflict. Therefore,
we need to delete this ID at the end (deleting /var/lib/consul) of the build process.

This fix also stops the service before deleting /var/lib/consul,
as most likely it restarts during the build phase before the instance is snapshoted
 and creates a new ID.